+++ WebApplication/src/store/modules/configuration.js
@@ -22,6 +22,7 @@ const state = () => ({
     requestedTagsForStudyList: [],
     hasExtendedFind: false,
     hasExtendedChanges: false,
+    advancedOptions: {}
 })
 
 ///////////////////////////// GETTERS
@@ -81,6 +82,9 @@ const mutations = {
     setCustomLogo(state, { customLogoUrl, hasCustomLogo }) {
         state.hasCustomLogo = hasCustomLogo;
         state.customLogoUrl = customLogoUrl;
+    },
+    setAdvancedOptions(state, { advancedOptions }) {
+        state.advancedOptions = advancedOptions;
     }
 
 }
@@ -118,6 +122,7 @@ const actions = {
         const oe2Config = await api.loadOe2Configuration();
         commit('setUiOptions', { uiOptions: oe2Config['UiOptions']});
         commit('setTokens', { tokens: oe2Config['Tokens']});
+        commit('setAdvancedOptions', { advancedOptions: oe2Config['AdvancedOptions']});
 
         if ('Profile' in oe2Config) {
             commit('setUserProfile', { profile: oe2Config['Profile']});